Split out classes under Abbyy module
